Classes

+++ csharp-basics/exercises/ClassesAndObjects/VideoStore/Program.cs
@@ -56,10 +56,8 @@ namespace VideoStore
             {
                 Console.WriteLine("Enter movie name");
                 string movieName = Console.ReadLine();
-
                 Console.WriteLine("Enter rating");
-                int rating = Convert.ToInt16(Console.ReadLine());
-
+                double rating = Convert.ToDouble(Console.ReadLine());
                 _videoStore.AddVideo(movieName);
                 _videoStore.TakeUsersRating(rating, movieName);
             }
